refactor(recorder): extract slot-load orchestration into slotSampleLoader
Move the cache -> maybe-download -> maybe-load policy out of the
SampleRecorder $effect into a plain component-free function that can be
reasoned about and tested without mounting the component. The effect now
only detects slot changes and delegates.
